requirements.txt
google-genai>=0.2.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
requests>=2.31.0
pydub>=0.25.1
python-dotenv>=1.0.0
